+++ R/apply-edit.R
@@ -6,19 +6,19 @@
 #' @param proj A `sondeproj` object holding sonde data.
 #' @param edit A list of length six:
 #' - data: new updated data as a `data.frame`
-#' - rows: logical vector which specifies rows changed as TRUE
+#' - rows: vector of indices that were changed
 #' - y_var: parameter being edited
 #' - step: name of the editing step for the changelog
 #' - note: an optional note to add to the changelog
 #' - flag: character flag to use for edits to the data
-#'
+#' @param username the username of the person who made the change
 #' @returns A `sondeproj` object with edits made.
 #' @export
 #' @md
 #' @examples
 #' data <- example_data
 #' data$fDOM_QSU[1:4] <- NA
-#' rows <- rep(FALSE, nrow(data))
+#' rows <- c(1:4)
 #' rows[1:4] <- TRUE
 #' edit <- list(data = example_data,
 #'              rows = rows,
@@ -26,13 +26,13 @@
 #'              step = "outlier removal",
 #'              note = "example edit",
 #'              flag = "RM07")
-#' updated_proj <- apply_edit(example_sondeproj, edit)
+#' updated_proj <- apply_edit(example_sondeproj, edit, "Smith")
 #'
-apply_edit <- function(proj, edit){
+apply_edit <- function(proj, edit, username){
   stopifnot(is.list(edit), inherits(proj, "sondeproj"))
 
   #skip applying edit if no rows changed
-  if(sum(edit$rows) == 0){
+  if(length(edit$rows) == 0){
     return(proj)
   }
 
@@ -42,7 +42,7 @@ apply_edit <- function(proj, edit){
 
   #apply flags to data preserving any existing flags
     #if all loop through
-    index <- which(edit$rows)
+    index <- edit$rows
     if(edit$y_var == "all"){
       y_vars <- get_parms(olddata)
       for(v in y_vars){newdata <- add_flags(newdata, v,index, edit$flag)}
@@ -55,8 +55,9 @@ apply_edit <- function(proj, edit){
     names(dif) <- diff_version(proj) #give name to list item
 
   #update log entry
-    proj <- write_log(proj, edit$y_var, edit$step, n=sum(edit$rows, na.rm=TRUE),
-                      note = edit$note, diff_name = names(dif), return = "sondeproj")
+    proj <- write_log(proj, edit$y_var, edit$step, n=length(edit$rows),
+                      note = edit$note, diff_name = names(dif), return = "sondeproj",
+                      user=username)
 
   #add in new df and diff
     proj$data <- newdata