+++ inst/app/app.R
@@ -36,10 +36,6 @@ ui <-  page_fillable(
                 value = "step-3",
                 SondePolishR::check_data_UI("data3")
                 ),
-      # nav_panel("Quality Flags",
-      #           value = "step-4",
-      #           SondePolishR::quality_UI("data4")
-      # ),
       nav_panel("Physical Limits",
                 value = "step-5",
                 SondePolishR::limits_UI("data5")
@@ -48,14 +44,14 @@ ui <-  page_fillable(
                 value = "step-6",
                 SondePolishR::outlier_UI("data6")
       ),
-      nav_panel("Interpolation",
-                value = "step-7",
-                SondePolishR::interp_UI("data7")
-      ),
       nav_panel("Corrections",
                 value = "step-8",
                 SondePolishR::correction_UI("data8")
                 ),
+      nav_panel("Interpolation",
+                value = "step-7",
+                SondePolishR::interp_UI("data7")
+      ),
       nav_panel("fDOM Corrections",
                 value = "step-9",
                 SondePolishR::fdom_UI("data9")
@@ -87,6 +83,7 @@ server <- function(input, output, session) {
     data_ver <- reactiveVal(0) #keeping track of when new data is uploaded
     y_var <- reactiveVal(NULL) #the y-variable being looked at
     current_mod <- reactiveVal() # the module being viewed
+    username <- reactiveVal(Sys.info()[["user"]]) #name to use for changelog, uses username by default
 
     #holds the viewing state to sync across the modules
     view_state <- reactiveVal(list(abs_dates =NULL,
@@ -101,31 +98,28 @@ server <- function(input, output, session) {
     })
 
   #step 1: load data
-   SondePolishR::load_data_server("data1", sondeproj, data_ver, view_state)
+   SondePolishR::load_data_server("data1", sondeproj, data_ver, view_state,username)
 
   #step 2: plot data
-   SondePolishR::explore_data_server("data2", sondeproj, data_ver, y_var, view_state)
+   SondePolishR::explore_data_server("data2", sondeproj, data_ver, y_var, view_state,username)
 
   #step 3: check data
-   SondePolishR::check_data_server("data3", sondeproj, data_ver, y_var)
+   SondePolishR::check_data_server("data3", sondeproj, data_ver, y_var,username)
 
-  # #step 4: quality checks
-  #  SondePolishR::quality_server("data4", sondeproj, data_ver, y_var, view_state)
-
-  #step 5: physical limits
-   SondePolishR::limits_server("data5", sondeproj, data_ver, y_var, view_state)
+   #step 5: physical limits
+   SondePolishR::limits_server("data5", sondeproj, data_ver, y_var, view_state,username)
 
   #step 6: outlier corrections
-   SondePolishR::outlier_server("data6", sondeproj, data_ver, y_var, view_state)
-
-  #step 7: data interpolation
-   SondePolishR::interp_server("data7", sondeproj, data_ver, y_var, view_state, current_mod)
+   SondePolishR::outlier_server("data6", sondeproj, data_ver, y_var, view_state,username)
 
   #step 8: additive shift
-   SondePolishR::correction_server("data8", sondeproj, data_ver, y_var, view_state)
+   SondePolishR::correction_server("data8", sondeproj, data_ver, y_var, view_state,username)
+
+  #step 7: data interpolation
+   SondePolishR::interp_server("data7", sondeproj, data_ver, y_var, view_state, username, current_mod)
 
   #step 9: fdom corrections
-   SondePolishR::fdom_server("data9", sondeproj, data_ver, y_var, view_state)
+   SondePolishR::fdom_server("data9", sondeproj, data_ver, y_var, view_state,username)
 
   #step 10: export data
    SondePolishR::export_server("data10", sondeproj, data_ver, y_var,current_mod)