+++ tests/testthat/test-step2-explore-plot.R
@@ -9,6 +9,7 @@ test_that("{shinytest2} recording: checking-module2", {
   app$upload_file(`data1-pj_file` = file.path(test_path(), "testdata", "example-sonde-project.RDS"))
 
   #click to load files and create project
+  app$set_inputs(`data1-username` = "Smith") #so my username isn't stored with tests
   app$click("data1-load_prj")
 
   #check initial plot is made
@@ -96,23 +97,17 @@ test_that("{shinytest2} recording: checking-module2", {
   app$wait_for_idle()
   plot_obj <- app$get_value(export = "data2-plot_obj")
   expect_snapshot_value(get_plotly_snap(plot_obj), style = "json2")
-  app$expect_screenshot(name = "change_variable")
+  app$expect_screenshot(name = "removing_OOW")
 
-  #check on the table
+#check other tables
+  app$set_inputs(`data2-table_opt` = "Change Log")
   app$expect_values(export = "data2-table", name="changelog-table",screenshot_args = FALSE)
 
-  #change to get other tables
   app$set_inputs(`data2-table_opt` = "Field Form")
-  app$expect_values(export = "data2-table", name="fieldform-table",screenshot_args = FALSE) #fieldform
+  app$expect_values(export = "data2-table", name="fieldform-table",screenshot_args = FALSE)
 
   app$set_inputs(`data2-table_opt` = "Calibration Check")
-  app$expect_values(export = "data2-table", name="calcheck-table",screenshot_args = FALSE) #cal check
-
-  app$set_inputs(`data2-table_opt` = "Data Summary")
-  app$expect_values(export = "data2-table", name="datasum-table2",screenshot_args = FALSE) #data summary
-
-  app$set_inputs(`data2-date_nav-period_view` = TRUE)
-  app$expect_values(export = "data2-table", name="datasum-table-weekly",screenshot_args = FALSE) #cal check
+  app$expect_values(export = "data2-table", name="calcheck-table",screenshot_args = FALSE)
 
   #test reverting changes (can't get to work)
   # app$set_inputs(`data2-date_nav-period_view` = FALSE)